+++ lib/output/constructor.go
@@ -198,7 +198,7 @@ type Config struct {
 	Resource           string                         `json:"resource" yaml:"resource"`
 	Retry              RetryConfig                    `json:"retry" yaml:"retry"`
 	S3                 writer.AmazonS3Config          `json:"s3" yaml:"s3"`
-	SFTP               writer.SFTPConfig              `json:"sftp" yaml:"sftp"`
+	SFTP               SFTPConfig                     `json:"sftp" yaml:"sftp"`
 	SNS                writer.SNSConfig               `json:"sns" yaml:"sns"`
 	SQL                SQLConfig                      `json:"sql" yaml:"sql"`
 	SQS                writer.AmazonSQSConfig         `json:"sqs" yaml:"sqs"`
@@ -266,7 +266,7 @@ func NewConfig() Config {
 		Resource:           "",
 		Retry:              NewRetryConfig(),
 		S3:                 writer.NewAmazonS3Config(),
-		SFTP:               writer.NewSFTPConfig(),
+		SFTP:               NewSFTPConfig(),
 		SNS:                writer.NewSNSConfig(),
 		SQL:                NewSQLConfig(),
 		SQS:                writer.NewAmazonSQSConfig(),